refactor: mapper - controller 분리 #238
